🧩 refactor(question): move QuestionDifficulty to domain layer
Fixes architecture violation where domain entity imported from infrastructure.
QuestionDifficulty is now a domain value object, ensuring clean architecture
dependency rules are followed (domain → infrastructure, not vice versa).

- Created QuestionDifficulty value object in domain/value-objects
- Updated all imports to use domain location
- Infrastructure now imports from domain (correct direction)
- Updated validation schemas to use domain constants
