+++ apps/api/src/features/question/infrastructure/drizzle/schema/enums.ts
@@ -1,5 +1,6 @@
 // Question feature specific enums
 import { pgEnum } from 'drizzle-orm/pg-core';
+import { QUESTION_DIFFICULTY_VALUES } from '../../../domain/value-objects/QuestionDifficulty';
 
 // Question related enums
 const questionTypeValues = ['single', 'multiple'] as const;
@@ -8,11 +9,11 @@ export const questionTypeEnum = pgEnum('question_type', questionTypeValues);
 const questionStatusValues = ['draft', 'active', 'inactive', 'archived'] as const;
 export const questionStatusEnum = pgEnum('question_status', questionStatusValues);
 
-// Difficulty is used in question context
-const difficultyValues = ['Beginner', 'Intermediate', 'Advanced', 'Expert', 'Mixed'] as const;
-export const difficultyEnum = pgEnum('difficulty', difficultyValues);
+// Difficulty enum using values from domain layer
+export const difficultyEnum = pgEnum('difficulty', QUESTION_DIFFICULTY_VALUES);
 
 // Type aliases for ergonomic usage
 export type QuestionType = (typeof questionTypeValues)[number];
 export type QuestionStatus = (typeof questionStatusValues)[number];
-export type QuestionDifficulty = (typeof difficultyValues)[number];
+// Re-export QuestionDifficulty from domain for backward compatibility
+export type { QuestionDifficulty } from '../../../domain/value-objects/QuestionDifficulty';
